Lower: implemented additional GenreServiceImplTest.

+++ src/main/java/hamdam/bookee/APIs/genre/GenreServiceImpl.java
@@ -2,8 +2,8 @@ package hamdam.bookee.APIs.genre;
 
 import hamdam.bookee.APIs.book.BookEntity;
 import hamdam.bookee.APIs.book.BookRepository;
-import hamdam.bookee.tools.exceptions.ResourceNotFoundException;
 import hamdam.bookee.tools.exceptions.ApiResponse;
+import hamdam.bookee.tools.exceptions.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.BeanUtils;
 import org.springframework.data.domain.Page;
@@ -24,8 +24,7 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public GenreDTO addGenre(GenreDTO dto) {
-        GenreEntity entity = new GenreEntity(dto);
-        genreRepository.save(entity);
+        genreRepository.save(new GenreEntity(dto));
         return dto;
     }
 
@@ -47,7 +46,6 @@ public class GenreServiceImpl implements GenreService {
                 -> new ResourceNotFoundException("Genre", "id", id));
         BeanUtils.copyProperties(genreDTO, oldGenre, "id");
 
-        // TODO: 9/2/22 do yo really need 3 genre related object: oldGenre, newGenre, genre(DTO)?
         // TODO: 9/2/22 why calling copyProperties?
         List<BookEntity> books = new ArrayList<>();
         genreDTO.getBooks().forEach(bookId -> {
@@ -63,7 +61,9 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public ApiResponse deleteGenre(Long id) {
-        genreRepository.existsById(id);
+        if (!genreRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Genre", "id", id);
+        }
         genreRepository.deleteById(id);
         return new ApiResponse(
                 HttpStatus.NO_CONTENT,
